requirements.txt
streamlit
Pillow
langdetect
google-cloud-vision
google-generativeai
python-dotenv
faiss-cpu
langchain
langchain-community
langchain-huggingface
sentence-transformers
requests
PyMuPDF
pandas
openpyxl
